pipline

+++ src/Controller/EventsController.php
@@ -3,7 +3,6 @@
 namespace App\Controller;
 
 use Doctrine\DBAL\Types\Types;
-
 use App\Entity\Events;
 use App\Form\EventType;
 use App\Repository\EventsRepository;
@@ -18,9 +17,6 @@ use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 
 class EventsController extends AbstractController
 {
-
-
-
     #[Route('/home', name: 'app_home')]
     public function home(): Response
     {
@@ -31,8 +27,8 @@ class EventsController extends AbstractController
 
     #[Route('/events', name: 'app_event')]
     public function listEvents(EventsRepository $er): Response
-
-    {   $this->denyAccessUnlessGranted('ROLE_ADMIN');
+    {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
         $events = $er->findAll();
         return $this->render('events/listEvents.html.twig', [
             'events' => $events,
@@ -52,7 +48,8 @@ class EventsController extends AbstractController
 
     #[Route('/events/filter', name: 'app_event_filter', methods: ['GET'])]
     public function filterEvents(Request $request, EventsRepository $er): Response
-    {    $this->denyAccessUnlessGranted('ROLE_ADMIN');
+    {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
         $field = $request->query->get('field');
         $value = $request->query->get('value');
 
@@ -70,7 +67,8 @@ class EventsController extends AbstractController
     }
     #[Route('/events/new', name: 'app_new')]
     public function new(Request $request, EntityManagerInterface $em): Response
-    {    $this->denyAccessUnlessGranted('ROLE_ADMIN');
+    {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
         $event = new Events();
 
         $form = $this->createForm(EventType::class, $event);
@@ -91,7 +89,8 @@ class EventsController extends AbstractController
 
     #[Route('/event/edit/{id}', name: 'app_event_edit')]
     public function edit(Events $event, Request $request, EntityManagerInterface $entityManager): Response
-    {    $this->denyAccessUnlessGranted('ROLE_ADMIN');
+    {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
         // Création du formulaire avec l'entité existante
         $form = $this->createForm(EventType::class, $event);
 
@@ -116,7 +115,8 @@ class EventsController extends AbstractController
     // Delete Event
     #[Route('/events/delete/{id}', name: 'app_event_delete', methods: ['POST'])]
     public function deleteEvent(Request $request, Events $event, EntityManagerInterface $em): Response
-    {    $this->denyAccessUnlessGranted('ROLE_ADMIN');
+    {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
         if ($this->isCsrfTokenValid('delete' . $event->getId(), $request->request->get('_token'))) {
             $em->remove($event);
             $em->flush();
@@ -124,7 +124,4 @@ class EventsController extends AbstractController
 
         return $this->redirectToRoute('app_event');
     }
-
-
-
 }
